Made zshrc more compatible

+++ zshrc
@@ -24,8 +24,14 @@ zstyle ':completion:*' completer _expand _complete _correct _approximate
 zstyle ':completion:*' format 'Completing %d'
 zstyle ':completion:*' group-name ''
 zstyle ':completion:*' menu select=2
-eval "$(dircolors -b)"
-zstyle ':completion:*:default' list-colors ${(s.:.)LS_COLORS}
+if whence dircolors >/dev/null; then
+  eval "$(dircolors -b)"
+  zstyle ':completion:*:default' list-colors ${(s.:.)LS_COLORS}
+  alias ls='ls --color'
+else
+  export CLICOLOR=1
+  zstyle ':completion:*:default' list-colors ''
+fi
 zstyle ':completion:*' list-colors ''
 zstyle ':completion:*' list-prompt %SAt %p: Hit TAB for more, or the character to insert%s
 zstyle ':completion:*' matcher-list '' 'm:{a-z}={A-Z}' 'm:{a-zA-Z}={A-Za-z}' 'r:|[._-]=* r:|=* l:|=*'
@@ -38,7 +44,7 @@ zstyle ":completion:*:commands" rehash 1
 zstyle ':completion:*:*:kill:*:processes' list-colors '=(#b) #([0-9]#)*=0=01;31'
 zstyle ':completion:*:kill:*' command 'ps -u $USER -o pid,%cpu,tty,cputime,cmd'
 
-. /home/igel/dev/files/prompt.zsh
+. $HOME/dev/files/prompt.zsh
 
 export STUDIO_JDK=/home/igel/dev/jdk1.7.0_25
 export REDIS_PATH=/usr/bin/redis-server
@@ -49,8 +55,10 @@ alias suspend='dbus-send --system --print-reply --dest="org.freedesktop.UPower" 
 
 export BROWSER=/usr/bin/firefox
 
-# Show message of the day in new consoles
-for i in /etc/update-motd.d/*; do if [ "$i" != "/etc/update-motd.d/98-fsck-at-reboot" ]; then $i 2>/dev/null; fi; done
+if [ -d /etc/update-motd.d ]; then
+  # Show message of the day in new consoles
+  for i in /etc/update-motd.d/*; do if [ "$i" != "/etc/update-motd.d/98-fsck-at-reboot" ]; then $i 2>/dev/null; fi; done
+fi
 
 function beep_on_return () (
   echo -ne '\a'
